moved add_reference to views directory
