+++ src/app.py
@@ -4,7 +4,6 @@ Defines the Flask applications dependencies and database connection.
 import os
 from flask import Flask, request, render_template
 from flask_sqlalchemy import SQLAlchemy
-from utils.reference_handler import create_reference
 
 
 if os.environ.get('ENVIRONMENT') == 'production':
@@ -27,26 +26,6 @@ def hello_world():
 
     return "<p>Hello, World!</p>"
 
-@app.route("/add_reference", methods=["GET", "POST"])
-def add_reference():
-    """
-    Route for handling adding a new reference
-    """
-    if request.method == "GET":
-        return render_template("add_reference.html")
-
-    if request.method == "POST":
-        author = request.form['author']
-        title = request.form['title']
-        year = request.form['year']
-
-        ref = create_reference(author, title, year)
-
-        return f"{ref.author}<br>{ref.title}<br>{ref.year}"
-
-        # Add processing for reference into database
-
-    return None
 
 @app.route("/view_reference")
 def view_reference():
@@ -57,6 +36,7 @@ def view_reference():
 
     return render_template("view_reference.html", references=reference_list)
 
+import views.add_reference
 
 if __name__ == "__main__":
     app.run(host='0.0.0.0')
